+++ crud-Js/README.md
@@ -2,4 +2,10 @@
 json-server --watch db.json
 ```
 
+- [x] [Time-saving synchronised browser testing.](https://browsersync.io/)
+
+```sh
+npm install -g browser-sync
+```
+
 Browser sync: browser-sync start --server --file . --host --port 5000 --startPath screens/lista_cliente.html
