+++ crud-Js/screens/lista_cliente.html
@@ -70,6 +70,6 @@
         </article>
       </div>
     </main>
-    <script src="../service/client-service.js"></script>
+    <script type="module" src="../controllers/client.controller.js"></script>
   </body>
 </html>
